Refactor typings
