+++ packages/cli-kit/src/public/node/typing/deep-required.ts
@@ -1,11 +1,11 @@
 /**
- * Converts a mapping type to be non-optional
+ * Converts a mapping type to be non-optional.
  *
- * ```
+ * ```ts
  * type T = DeepRequired<{optionalKey?: string, nullableValue: string | null, undefinableValue: string | undefined}>
  * T = {optionalKey: string, nullableValue: string, undefinableValue: string}
  * ```
- *
+ *.
  */
 export type DeepRequired<T> = {
   [TKey in keyof Required<T>]: NonNullable<Required<T>[TKey]>
